10_14

+++ src/main/java/com/server/youtube/service/CommentService.java
@@ -32,9 +32,43 @@ public class CommentService {
     }
     //댓글 삭제 하기
     public void remove(int commentCode){
-        dao.deleteById(commentCode);
-    }
+        
+        // 조건 : 자식 댓글이 있는지 체크
+        //  SELECT count(*) FROM comment WHERE parent_code = :commentCode
+        List<Comment> comments = queryFactory.selectFrom(qComment)
+                .where(qComment.parentCode.eq(commentCode)).fetch();
+        int childCount = comments.size();
 
+        Comment comment = dao.findById(commentCode).get();
+
+        if(childCount > 0){
+            // 자식 댓글이 있는 경우
+            // -> 부모 댓글은 삭제되지 않고 "삭제된 댓글입니다" 표시
+            // -> isDelete 값 true로 변환
+            comment.setDelete(true);
+            dao.save(comment);
+        }else{
+            // 자식 댓글이 없는 경우
+            // -> 부모 댓글 DB 에서 완전 삭제
+            dao.deleteById(commentCode);
+        }
+        // 삭제된 부모 댓글인지 확인
+        // 해당 댓글의 부모 댓글이 있는지 체크
+        if(comment.getParentCode() > 0){
+            // 부모 댓글의 자식 댓글이 모두 삭제되었는지 체크
+            List<Comment> parents = queryFactory.selectFrom(qComment)
+                    .where(qComment.parentCode.eq(comment.getParentCode())).fetch();
+            int parentCount = parents.size(); // 0인 경우
+            if(parentCount == 0) {
+                // 부모 댓글이 "삭제된 댓글입니다" isDelete가 true인 상태인 경우
+                Comment parent = dao.findById(comment.getParentCode()).get();
+                if(parent.isDelete()){
+                    dao.deleteById(parent.getCommentCode());
+                }
+            }
+        }
+    }
+    
 
 
     // 비디오별 상위 댓글들 보여주기 -> SQL문 짜보기!
